requirements.txt
pycryptodome
Flask
flask_cors
requests

